requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Werkzeug==3.0.1
python-dotenv==1.0.0
Pillow==10.1.0
Markdown==3.5.1
bleach==6.1.0
Pygments==2.17.2
PyJWT==2.8.0
slugify
email-validator==2.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
pytest==8.2.0
qrcode==7.4.2
